requirements.txt
# The app itself
-e .

Django>=1.11,<2.2
mock
psycopg2-binary
file-magic

# example site
django-recaptcha
